refactor: Overweighted views and multilines if
